% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proposals_info.R
\name{get_proposals_info}
\alias{get_proposals_info}
\title{Scrape stats from tagging proposals}
\usage{
get_proposals_info(urls, verbose = FALSE)
}
\arguments{
\item{urls}{a string containing the tagging proposal to retrieve details from.}

\item{verbose}{a boolean to print the URL in the console.}
}
\value{
a dataframe with the following columns:
\itemize{
\item \code{url}: the URL of the tagging proposal.
\item \code{page_creator}: the username who created the page.
\item \code{date_of_page_creation}: the date in which the page was created.
\item \code{latest_editor}: the username who last edited the page.
\item \code{date_of_latest_edit}: date in which the page was last updated.
\item \code{total_number_of_edits}: number of total times that the page has been edited.
\item \code{total_number_of_disctinct_authors}: number of distinct users who have edited the page.
}
}
\description{
Given a vector of urls, scrapes the tagging proposals to retrieve statistics and stores them into a dataframe.
}
\examples{
proposals_info <- get_proposals_info('https://wiki.openstreetmap.org/wiki/Proposal:Electricity')

proposals_info

}
