% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_voting_summary.R
\name{get_voting_summary}
\alias{get_voting_summary}
\title{Scrape voting summary from tagging proposals}
\usage{
get_voting_summary(urls)
}
\arguments{
\item{urls}{a string containing the tagging proposal to retrieve details from.}
}
\value{
a dataframe with some stats
}
\description{
Given a vector of urls, scrapes the tagging proposals to retrieve voting history data.
}
\examples{
voting_summary <- get_voting_summary('https://wiki.openstreetmap.org/wiki/Proposal:Electricity')

voting_summary

}
