% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proposals}
\alias{proposals}
\title{OSM Tagging Proposals}
\format{
A data frame with 5 rows and 25
columns:
\describe{
\item{id}{A unique identifier for every boundary}
\item{service}{URL pointing to the API service}
\item{url_download}{URL querying the API service to return all features as a geojson file}

}
}
\source{
\url{https://wiki.openstreetmap.org/wiki/Proposal_process}
}
\usage{
proposals
}
\description{
Metadata tagging proposals in OSM Wiki ().
This dataset has been created by the script {data-raw/tagging_proposals.R}
}
\keyword{datasets}
