% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proposals_votes}
\alias{proposals_votes}
\title{OSM Tagging Proposals' votes Data}
\format{
columns:
\describe{
\item{url}{(string) the URL to the tagging proposal wiki page.}
\item{votes_raw}{a string with the wiki page containing the proposal.}
\item{vote}{(factor) vote type (approve, abstain, oppose).}
\item{user}{(factor) voter's username}

}
}
\source{
\url{https://wiki.openstreetmap.org/wiki/Proposal_process#Voting}
}
\usage{
proposals_votes
}
\description{
A data frame with 8487 rows and 4 columns, containing all
the votes cast for every tagging proposal in \code{proposals}, generated by \code{get_proposals_votes()}.
}
\keyword{datasets}
