% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proposals_info.R
\name{get_proposals_info}
\alias{get_proposals_info}
\title{Scrape stats from tagging proposals}
\usage{
get_proposals_info(urls)
}
\arguments{
\item{urls}{a string containing the tagging proposal to retrieve details from.}
}
\value{
a dataframe with some stats
}
\description{
Given a vector of urls, scrapes the tagging proposals to retrieve statistics and stores them into a dataframe.
}
\examples{
get_proposals_info('https://wiki.openstreetmap.org/wiki/Proposal:Electricity')

}
