% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tagging_proposals.R
\name{get_tagging_proposals}
\alias{get_tagging_proposals}
\title{Retrieve a list of tagging proposals}
\usage{
get_tagging_proposals(
  statuses = c("Rejected", "Approved", "Proposed"),
  details = TRUE,
  max_items = 5,
  info = TRUE,
  voting_summary = TRUE,
  file = NULL
)
}
\arguments{
\item{statuses}{a string or vector containing proposal's statuses to look for.}

\item{details}{boolean. Specifies whether to retrieve or not pages' details.}

\item{max_items}{a integer specifiying the the maximum number of members to retrieve for each category. Set to 5 by default.}

\item{info}{a boolean. Specifies whether to retrieve or not proposals' stats. Set TRUE by default.}

\item{voting_summary}{a boolean. Specifies whether to retrieve or not proposals' voting summary. Set TRUE by default.}

\item{file}{a string containing the file path to store the dataset.}
}
\value{
a dataframe containing tagging proposals and associated metadata (see \href{proposals.html}{\code{proposals}} for a description of its metadata).
}
\description{
Retrieves all tagging proposals in OSM's wiki by querying mediawiki's API and provides basic metadata.
}
\examples{
proposals <- get_tagging_proposals("Proposed")
head(proposals)

}
