% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_voting_summary.R
\name{get_voting_summary}
\alias{get_voting_summary}
\title{Scrape voting summary from tagging proposals}
\usage{
get_voting_summary(urls)
}
\arguments{
\item{urls}{a string containing the tagging proposal to retrieve details from.}
}
\value{
a dataframe with the following columns:
\itemize{
\item \code{url}: the URL of the tagging proposal.
\item \code{proposal_status}: the status of the proposal.
\item \code{proposed_by}: name of the user(s) making the proposal.
\item \code{tagging}: proposed tagging scheme.
\item \code{applies_to_node}: (Boolean) whether this tag applies to nodes or not.
\item \code{applies_to_way}: (Boolean) whether this tag applies to ways or not.
\item \code{applies_to_area}: (Boolean) whether this tag applies to areas or not.
\item \code{applies_to_relation}: (Boolean) whether this tag applies to relations or not.
\item \code{definition}: a short definition of what the tagging proposal aims to describe.
\item \code{statistics}: usage statistics, from \href{https://taginfo.geofabrik.de/}{taginfo}.
\item \code{rendered_as}: whether the feature has an icon or not.
\item \code{draft_started}: date in which the proposal draft started.
\item \code{rfc_start}: date in which Request For Comments (RFC) started.
\item \code{vote_start}: date in which the voting process started.
\item \code{vote_end}: date in which the voting process ended.
}
}
\description{
Given a vector of urls, scrapes the tagging proposals to retrieve voting history data.
}
\examples{
voting_summary <- get_voting_summary('https://wiki.openstreetmap.org/wiki/Proposal:Electricity')

voting_summary

}
